requirements.txt
Babel
Beaker
Chameleon
Jinja2
Mako
Markdown
MarkupSafe
PasteDeploy
PyICU
PyYAML
Pygments
SQLAlchemy
Sphinx
WebOb
alembic
colander==1.0a5
crypto
deform==0.9.9
docutils
ezodf
ipdb
ipython
lxml
networkx
nose
openpyxl
passlib
peppercorn
prettytable
psycopg2
pyasn1
pycrypto
pyinotify
pyparsing
pyramid
pyramid-beaker
pyramid-debugtoolbar
pyramid-jinja2
pyramid-mailer
pyramid-mako
pyramid-tm
python-dateutil
python-magic
pytz
regex
repoze.lru
repoze.sendmail
requests
rsa
six
sqlparse
tornado
transaction
translationstring
venusian
waitress
xlrd
yolk3k
zope.deprecation
zope.interface
zope.sqlalchemy
gunicorn
